Adds more modules
