+++ js/helpers/create-array-by-length.js
@@ -1,0 +1,3 @@
+export const createArrayByLength = function (arrayLength, arrayElement) {
+  return Array.from({length:arrayLength}, arrayElement);
+};
