+++ js/helpers/get-random-array-element.js
@@ -1,0 +1,3 @@
+import {getRandomInteger} from './get-random-integer.js';
+
+export const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
