+++ js/helpers/id-creators.js
@@ -1,3 +1,5 @@
+import {getRandomInteger} from './get-random-integer.js';
+
 export const createId = function () {
   let lastGeneratedId = 0;
 
@@ -5,14 +7,6 @@ export const createId = function () {
     lastGeneratedId += 1;
     return lastGeneratedId;
   };
-};
-
-export const getRandomInteger = function (min, max) {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
 };
 
 export const createRandomId = function (min, max) {
@@ -30,5 +24,3 @@ export const createRandomId = function (min, max) {
     return currentValue;
   };
 };
-
-export const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
